+++ changelog.md
@@ -1,4 +1,4 @@
-# Version 4.0.0 - dd.mm.2023
+# Version 4.0.0 - 13.09.2023
 
 ## Breaking changes